refactor(types): tabs
